// src/app/models/NewSupplier.ts
export class NewSupplier {
  id?: number;
  companyName?: string;
  address?: string;
  email?: string;
  legalPaperwork?: string;
  infoPaperwork?: string;
  addedDate?: string;
  password?: string;
  companyCode?: string;
  managerName?: string;
  managerPhone?: string;
  phone?: string;
  fax?: string;
  subCategoryID?: number;
}